Replace System Rules with System Lambda
System Lambda is more specific. It only wraps the part of the code that
writes to System.out. In addition System Lambda is independent from the
test framework and no obstacle for moving to another test framework
(e.g. JUnit Lambda).

The assertions have been changed from "contains" to "isEqualTo" because
SystemLambda's "tapSystemOutNormalized" takes care about normalizing the
new line chars so that the returned String is the same on all OS.